add buildClient using esbuild with html plugin, refactor App to receive lazy loaded report clients, fix loose client-ui exports
