+++ packages/geoprocessing/package.json
@@ -228,9 +228,11 @@
         "zx": "^4.3.0"
     },
     "devDependencies": {
+        "@craftamap/esbuild-plugin-html": "^0.6.1",
         "@testing-library/jest-dom": "^6.4.2",
         "@types/finalhandler": "^1.2.0",
         "esbuild": "0.20.2",
+        "esbuild-plugin-inline-image": "^0.0.9",
         "eslint": "^8.57.0",
         "eslint-plugin-unicorn": "^51.0.1",
         "identity-obj-proxy": "^3.0.0",
