+++ packages/geoprocessing/src/datasources/helpers.ts
@@ -23,7 +23,6 @@ import {
   RasterDatasource,
 } from "../types/index.js";
 import { DataClass } from "../types/index.js";
-import path from "path";
 
 /**
  * Returns the first item that returns true for filter
@@ -245,16 +244,4 @@ export function getDatasetBucketName<C extends BaseImportDatasourceConfig>(
   config: C
 ) {
   return `gp-${config.package.name}-datasets`;
-}
-
-export function getJsonPath(dstPath: string, datasourceId: string) {
-  return path.join(dstPath, datasourceId) + ".json";
-}
-
-export function getFlatGeobufPath(dstPath: string, datasourceId: string) {
-  return path.join(dstPath, datasourceId) + ".fgb";
-}
-
-export function getGeopackagePath(dstPath: string, datasourceId: string) {
-  return path.join(dstPath, datasourceId) + ".gpkg";
 }
