+++ packages/geoprocessing/src/index.ts
@@ -16,5 +16,6 @@ export * from "./types/index.js";
 export * from "./testing/index.js";
 export * from "./project/index.js";
 export * from "./util/index.js";
+export * from "./context/index.js"
 
 export { default as version } from "../package.json"  assert { type: "json" };
